requirements.txt
# Existing packages
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings
# CORS middleware
fastapi-cors==0.0.6

# JWT and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
firebase-admin

# Environment variables
python-dotenv==1.0.0

# Google APIs 
google-generativeai
googlemaps==4.10.0

# Logging and utilities
python-json-logger==2.0.7

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1


pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI endpoints

gunicorn==21.2.0
